requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==4.0.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
